textbook_analyzer/requirements.txt
# 基础依赖
pandas>=1.3.0
tqdm>=4.62.0
openpyxl>=3.0.9
python-dotenv>=0.19.0

# LLM依赖
openai>=1.0.0

# 文件处理
PyPDF2>=2.0.0
PyMuPDF>=1.19.0
python-docx>=0.8.11

# 其他工具
matplotlib>=3.5.0
seaborn>=0.11.2
numpy>=1.21.0 